+++ src/ui/CommitList.h
@@ -103,7 +103,6 @@ public:
   void setModel(QAbstractItemModel *model) override;
 
 signals:
-  void firstPainted();
   void statusChanged(bool dirty);
   void statusError(const QString &error);
   void selectedRangeChanged(const QString &range);
@@ -114,7 +113,6 @@ signals:
                       bool spontaneous = false);
 
 protected:
-  void paintEvent(QPaintEvent *event) override;
   void contextMenuEvent(QContextMenuEvent *event) override;
   void mouseMoveEvent(QMouseEvent *event) override;
   void mousePressEvent(QMouseEvent *event) override;
@@ -177,7 +175,6 @@ private:
   int mHistoryPrefetchTarget = 0;
 
   bool mRestoreSelection{true};
-  bool mFirstPainted = false;
   bool mPreserveSelectionDetails = false;
   bool mSuppressSelectionNotification = false;
   QString mViewportCommit;